app/pyproject.toml
[project]
name = "csv-ans-support-api"
version = "0.1.0"
description = "Офлайн RAG API для первой линии поддержки ЦСВ АНС"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "httpx==0.27.2",
  "qdrant-client==1.11.3",
  "sentence-transformers==3.2.0",
  "python-multipart==0.0.12",
  "python-docx==1.1.2",
  "pypdf==5.0.1",
  "prometheus-client==0.21.0",
  "psycopg[binary]==3.2.3"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
